# Fix for quadratic batching in #99