@tclements 2022-01-12
Just tested, my fix will fail on 3D arrays due to using hcat..coming up with a solution for that now.