@tclements 2022-01-12
Fast N-d array concatenation (for N > 2) seems to be a common problem that needs to be solved in Base https://github.com/JuliaLang/julia/issues/21672. Might need to table this PR until that is implemented or use an explicit solution for 3D arrays.