@CarloLucibello 2022-01-13
I think we can introduce something like this:
```julia
cat_fatures(xs::AbstractVector{<:AbstractVector}) = reduce(vcat, xs)
cat_fatures(xs::AbstractVector{<:AbstractMatrix}) = reduce(hcat, xs)
cat_fatures(xs::AbstractVector{<:AbstractArray{T,N}}) where {T,N} = reduce((x1, x2) -> cat(x1, x2, dims=N), xs)
```